ref: move split_trees to bulloak-foundry
