+++ crates/core/src/utils.rs
@@ -1,21 +1,19 @@
 use unicode_xid::UnicodeXID;
 
-use crate::constants::TREES_SEPARATOR;
-
 pub fn capitalize_first_letter(s: &str) -> String {
     let mut c = s.chars();
-    match c.next() {
-        None => String::new(),
-        Some(f) => f.to_uppercase().collect::<String>() + c.as_str(),
-    }
+    c.next()
+        .map(char::to_uppercase)
+        .map(|first| first.to_string() + c.as_str())
+        .unwrap_or_default()
 }
 
 pub fn lower_first_letter(s: &str) -> String {
     let mut c = s.chars();
-    match c.next() {
-        None => String::new(),
-        Some(f) => f.to_lowercase().collect::<String>() + c.as_str(),
-    }
+    c.next()
+        .map(char::to_lowercase)
+        .map(|first| first.to_string() + c.as_str())
+        .unwrap_or_default()
 }
 
 /// This function makes the appropriate changes to a string to
@@ -51,66 +49,14 @@ pub fn pluralize<'a>(
     }
 }
 
-/// Splits the input text into distinct trees, delimited by two consecutive
-/// newlines.
-pub fn split_trees(text: &str) -> Box<dyn Iterator<Item = &str> + '_> {
-    if text.trim().is_empty() {
-        return Box::new(std::iter::once(""));
-    }
-
-    let trees = text.split(TREES_SEPARATOR).map(str::trim);
-    let non_empty_trees = trees.filter(|s| !s.is_empty());
-    let no_isolated_comments =
-        non_empty_trees.filter(|tree| !only_comments(tree));
-
-    Box::new(no_isolated_comments)
-}
-
-fn only_comments(tree: &str) -> bool {
-    tree.lines().all(|l| l.trim().starts_with("//"))
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{split_trees, to_pascal_case};
+    use super::to_pascal_case;
 
     #[test]
     fn to_modifier() {
         assert_eq!(to_pascal_case("when only owner"), "WhenOnlyOwner");
         assert_eq!(to_pascal_case("when"), "When");
         assert_eq!(to_pascal_case(""), "");
-    }
-
-    #[test]
-    fn splits_trees() {
-        let test_cases = vec![
-            ("Foo_Test\n└── when something bad happens\n   └── it should revert", vec![
-                "Foo_Test\n└── when something bad happens\n   └── it should revert",
-            ]),
-            ("Foo_Test\n└── when something bad happens\n   └── it should revert\n\nFoo_Test2\n└── when something bad happens\n   └── it should revert", vec![
-                "Foo_Test\n└── when something bad happens\n   └── it should revert",
-                "Foo_Test2\n└── when something bad happens\n   └── it should revert",
-            ]),
-            // Test with varying numbers of newlines between tree splits
-            // Assumes behavior is the same for 2 or more newlines
-            ("Foo_Test\n└── when something bad happens\n   └── it should revert\n\n\nFoo_Test2\n└── when something bad happens\n   └── it should revert", vec![
-                "Foo_Test\n└── when something bad happens\n   └── it should revert",
-                "Foo_Test2\n└── when something bad happens\n   └── it should revert",
-            ]),
-            ("Foo_Test\n└── when something bad happens\n   └── it should revert\n\n\n\nFoo_Test2\n└── when something bad happens\n   └── it should revert", vec![
-                "Foo_Test\n└── when something bad happens\n   └── it should revert",
-                "Foo_Test2\n└── when something bad happens\n   └── it should revert",
-            ]),
-            ("Foo_Test\n└── when something bad happens\n   └── it should revert\n\n\n\n\nFoo_Test2\n└── when something bad happens\n   └── it should revert", vec![
-                "Foo_Test\n└── when something bad happens\n   └── it should revert",
-                "Foo_Test2\n└── when something bad happens\n   └── it should revert",
-            ]),
-        ];
-
-        for (input, expected) in test_cases {
-            let trees = split_trees(input);
-            let results: Vec<_> = trees.collect();
-            assert_eq!(results, expected, "Failed on input: {}", input);
-        }
     }
 }
